Brewfolio-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
beautifulsoup4==4.12.3
google-genai>=1.0.0
openai>=1.0.0
supabase==2.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
redis==5.1.0
upstash-redis==1.1.0
tenacity==9.0.0
pytest==8.3.3
pytest-asyncio==0.24.0
